+++ config/routes.rb
@@ -1,4 +1,7 @@
 Rails.application.routes.draw do
   get '/analyses' => 'analyses#manage'
-  post '/analyses/learn' => 'analyses#learn'
+  post '/analyses' => 'analyses#execute'
+
+  get '/predictions' => 'predictions#manage'
+  post '/predictions' => 'predictions#execute'
 end