+++ spec/requests/manage_predictions_spec.rb
@@ -1,0 +1,16 @@
+# coding: utf-8
+require 'rails_helper'
+
+describe 'ブラウザで予測する', :type => :request do
+  include_context 'WebDriverを起動してCookieをセットする'
+
+  describe '予測画面を開く' do
+    before(:all) { @driver.get("#{base_url}/predictions") }
+
+    it '予測画面が表示されていること' do
+      is_asserted_by { @driver.current_url == "#{base_url}/predictions" }
+      is_asserted_by { @driver.find_element(:id, 'prediction_model') }
+      is_asserted_by { @driver.find_element(:xpath, '//form/input[@value="実行"]') }
+    end
+  end
+end